clrCall {rClr}	R Documentation

Call a method on an object

Description

Call a method on an object

Usage

clrCall(obj, methodName, ...)

Arguments

obj
an object

methodName
the name of a method of the object

...
additional method arguments passed to .External

Value

an object resulting from the call. May be a CLR object, or a native R object for common types. Can be NULL.

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
(testObj <- clrNew(testClassName))
clrCall(testObj, 'GetFieldIntegerOne')
## derived from unit test for matching the right method (function) to call.
f <- function(...){ paste( 'This called a method with arguments:', 
  paste(clrCallStatic('Rclr.TestMethodBinding', 'SomeStaticMethod', ...), collapse=', ')) }
f(1:3)
f(3)
f('a')
f('a', 3)
f(3, 'a')
f(list('a', 3))

## End(Not run)

[Package rClr version 0.9.1 Index]