clrCallStatic {rClr}	R Documentation

Call a static method on a CLR type

Description

Call a static method on a CLR type

Usage

clrCallStatic(typename, methodName, ...)

Arguments

typename
type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

methodName
the name of a static method of the type

...
additional method arguments passed to .External

Value

an object resulting from the call. May be a CLR object, or a native R object for common types. Can be NULL.

Examples

## Not run: 
library(rClr)
cTypename <- "Rclr.TestCases"
clrCallStatic(cTypename, "IsTrue", TRUE)

## End(Not run)

[Package rClr version 0.9.1 Index]