clrGet {rClr}	R Documentation

Gets the value of a field or property of an object or class

Description

Gets the value of a field or property of an object or class

Usage

clrGet(objOrType, name)

Arguments

objOrType
a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

name
the name of a field/property of the object

Value

an object resulting from the call. May be a CLR object, or a native R object for common types. Can be NULL.

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrReflect(testObj)
clrGet(testObj, 'FieldIntegerOne')
clrGet(testClassName, 'StaticPropertyIntegerOne')

## End(Not run)

[Package rClr version 0.9.1 Index]