clrGetEnumNames {rClr}	R Documentation

Gets the names of a CLR Enum value type

Description

Gets the names of a CLR Enum value type

Usage

clrGetEnumNames(enumType)

Arguments

enumType
a CLR object, System.Type or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

Value

a character vector of the names for the enum

Examples

## Not run: 
library(rClr)
clrGetEnumNames('Rclr.TestEnum')

## End(Not run)

[Package rClr version 0.9.1 Index]