clrGetInnerPkgName {rClr}	R Documentation

Gets the inner name used for the package

Description

Gets the inner name used for the package (rClrMono or rClrMs). This is not intented for use by most users.

Usage

clrGetInnerPkgName()

Value

the short name of the library currently loaded, depending on the runtime used (Mono or Microsoft .NET)

[Package rClr version 0.9.1 Index]