clrGetLoadedAssemblies {rClr}	R Documentation

List the names of loaded CLR assemblies

Description

List the names of loaded CLR assemblies

Usage

clrGetLoadedAssemblies(fullname = FALSE, filenames = FALSE)

Arguments

fullname
should the full name of the assemblies be returned

filenames
if TRUE, return a data frame where the second column is the URI (usually file path) of the loaded assembly.

Value

the names of loaded CLR assemblies

Examples

## Not run: 
library(rClr)
clrGetLoadedAssemblies()

## End(Not run)

[Package rClr version 0.9.1 Index]