clrGetMemberSignature {rClr}	R Documentation

Gets the signature of a CLI object member

Description

Gets a string representation of the signature of a member (i.e. field, property, method). Mostly used to interactively search for what arguments to pass to a method.

Usage

clrGetMemberSignature(clrobj, memberName)

Arguments

clrobj
CLR object

memberName
The exact name of the member (i.e. field, property, method) to search for

Value

a character vector with summary information on the method/member signatures

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrReflect(testObj)
clrGetMemberSignature(testObj, 'set_PropertyIntegerOne')
clrGetMemberSignature(testObj, 'FieldIntegerOne')
clrGetMemberSignature(testObj, 'PropertyIntegerTwo')

## End(Not run)

[Package rClr version 0.9.1 Index]