clrGetMethods {rClr}	R Documentation

List the instance methods of a CLR object

Description

List the instance methods of a CLR object

Usage

clrGetMethods(clrobj, contains = "")

Arguments

clrobj
CLR object

contains
a string that the methods names returned must contain

Value

a list of names of the methods of the CLR object

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrGetMethods(testObj)
clrGetMethods(testObj, 'Get')

## End(Not run)

[Package rClr version 0.9.1 Index]