clrGetNativeLibName {rClr}	R Documentation

Gets the name of the native library currently loaded.

Description

Gets the name of the native library currently loaded. Used only for unit tests.

Usage

clrGetNativeLibName()

Value

the name of the native library currently loaded: rClrMs or rClrMono

[Package rClr version 0.9.1 Index]