clrGetProperties {rClr}	R Documentation

List the instance properties of a CLR object

Description

List the instance properties of a CLR object

Usage

clrGetProperties(clrobj, contains = "")

Arguments

clrobj
CLR object

contains
a string that the property names returned must contain

Value

a list of names of the properties of the CLR object

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
testObj <- clrNew(testClassName)
clrGetProperties(testObj)
clrGetProperties(testObj, 'One')

## End(Not run)

[Package rClr version 0.9.1 Index]