clrGetStaticMemberSignature {rClr}	R Documentation

Gets the signature of a static member of a type

Description

Gets the signature of a static member of a type

Usage

clrGetStaticMemberSignature(typename, memberName)

Arguments

typename
type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

memberName
The exact name of the member (i.e. field, property, method) to search for

[Package rClr version 0.9.1 Index]