clrGetType {rClr}	R Documentation

Gets the type of a CLR object given its type name

Description

Gets the type of a CLR object given its type name

Usage

clrGetType(objOrTypename)

Arguments

objOrTypename
a character vector of length one. It can be the full file name of the assembly to load, or a fully qualified assembly name, or as a last resort a partial name.

Value

the CLR Type.

[Package rClr version 0.9.1 Index]