clrIs {rClr}	R Documentation

Check whether an object is of a certain type

Description

Check whether an object is of a certain type. This function is meant to match the behavior of the 'is' keyword in C#.

Usage

clrIs(obj, type)

Arguments

obj
an object

type
the object type to check for. It can be a character, of a object of CLR type System.RuntimeType

Value

TRUE or FALSE

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
(testObj <- clrNew(testClassName))
clrIs(testObj, testClassName)
clrIs(testObj, 'System.Object')
clrIs(testObj, 'System.Double')
(testObj <- clrNew('Rclr.TestMethodBinding'))
# Test for interface definitions
clrIs(testObj, 'Rclr.ITestMethodBindings')
clrIs(testObj, clrGetType('Rclr.ITestMethodBindings'))
clrIs(testObj, clrGetType('Rclr.TestMethodBinding'))
clrIs(testObj, clrGetType('System.Reflection.Assembly'))

## End(Not run)

[Package rClr version 0.9.1 Index]