clrLoadAssembly {rClr}	R Documentation

Loads a Common Language assembly.

Description

Loads an assembly. Note that this is loaded in the single application domain that is created by rClr, not a separate application domain.

Usage

clrLoadAssembly(name)

Arguments

name
a character vector of length one. It can be the full file name of the assembly to load, or a fully qualified assembly name, or as a last resort a partial name.

See Also

.C which this function wraps

Examples

## Not run: 
library(rClr)
clrGetLoadedAssemblies()
f <- file.path('SomeDirectory', 'YourDotNetBinaryFile.dll')
f <- path.expand(f)
stopifnot( file.exists(f) )
clrLoadAssembly(f)
# Load an assembly from the global assembly cache (GAC)
clrLoadAssembly("System.Windows.Presentation, 
  Version=3.5.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")
# The use of partial assembly names is discouraged; nevertheless it is supported
clrLoadAssembly("System.Web.Services")
clrGetLoadedAssemblies()

## End(Not run)

[Package rClr version 0.9.1 Index]