clrNew {rClr}	R Documentation

Create a new CLR object

Description

Create a new CLR object

Usage

clrNew(typename, ...)

Arguments

typename
type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

...
additional method arguments passed to the object constructor via the call to .External

Value

a CLR object

Examples

## Not run: 
library(rClr)
testClassName <- "Rclr.TestObject";
(testObj <- clrNew(testClassName))
# object with a constructor that has parameters
(testObj <- clrNew(testClassName, as.integer(123)))
clrLoadAssembly("System.Windows.Forms, Version=2.0.0.0, 
  Culture=neutral, PublicKeyToken=b77a5c561934e089")
f <- clrNew('System.Windows.Forms.Form')
clrSet(f, 'Text', "Hello from '.NET'")
clrCall(f, 'Show')

## End(Not run)

[Package rClr version 0.9.1 Index]