clrSetEnumProperty {rClr}	R Documentation

Sets the value of an enum field or property of an object or class

Description

Sets the value of an enum field or property of an object or class

Usage

clrSetEnumProperty(objOrType, name, enumtype, enumval)

Arguments

objOrType
a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

name
the name of a field/property of the object

enumtype
type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

enumval
the value to set the field with

[Package rClr version 0.9.1 Index]