clrToString {rClr}	R Documentation

Calls the ToString method of an object

Description

Calls the ToString method of an object as represented in the CLR. This function is here to help quickly test object equivalence from the R interpreter, for instance on the tricky topic of date-time conversions

Usage

clrToString(x)

Arguments

x
any R object, which is converted to a CLR object on which to call ToString

Value

the string representation of the object in the CLR

Examples

## Not run: 
library(rClr)
dt <- as.POSIXct('2001-01-01 02:03:04', tz='UTC')
clrToString(dt)

## End(Not run)

[Package rClr version 0.9.1 Index]