clrVT {rClr}	R Documentation

Get the COM variant type of a CLR object

Description

Get the COM variant type of a CLR object, e.g. "VT_ARRAY | VT_I8". This function only works when run on the Microsoft implementation of the CLR. This function is useful for advanced diagnosis; most users should never have to use it.

Usage

clrVT(objOrType, methodName, ...)

Arguments

objOrType
a CLR object, or type name, possibly namespace and assembly qualified type name, e.g. 'My.Namespace.MyClass,MyAssemblyName'.

methodName
the name of the method called on the object or class specified with objOrType

...
one or more arguments to pass to the function call

Value

A string

Examples

## Not run: 
library(rClr)
cTypename <- "Rclr.TestCases"
#         public static bool IsTrue(bool arg)
clrVT(cTypename, 'IsTrue', TRUE)
clrVT('System.Convert', 'ToInt64', 123L)
clrVT('System.Convert', 'ToUInt64', 123L)

## End(Not run)

[Package rClr version 0.9.1 Index]