dotOnLoad {rClr}	R Documentation

rClr .onLoad

Description

Function called when loading the rClr package with 'library'.

Usage

.onLoad(libname = "~/R", pkgname = "rClr")

Arguments

libname
the path to the library from which the package is loaded

pkgname
the name of the package.

Details

The function looks by default for the rClr native library for the Mono runtime. If the platform is Linux, this is the only option. If the platform is Windows, using the Microsoft .NET runtime is an option. If the rClr native library for MS.NET is detected, the Microsoft .NET runtime is loaded in preference to Mono.

[Package rClr version 0.9.1 Index]