getClrVersionString {rClr}	R Documentation

Gets the version of the common language runtime in use

Description

Gets the version of the common language runtime in use.

Usage

getClrVersionString()

Value

the version of the common language runtime in use

[Package rClr version 0.9.1 Index]