getCurrentConvertedObject {rClr}	R Documentation

System function to get a direct access to an object

Description

This function needs to be exported, but is highly unlikely to be of any use to an end user, even an advanced one. This is indirectly needed to unlock the benefits of using R.NET convert data structures between R and .NET. Using this function is a critical part of solving the rather complicated issue rClr #33.

Usage

getCurrentConvertedObject()

Value

a CLR object

[Package rClr version 0.9.1 Index]