getNativeLibsPath {rClr}	R Documentation

Architecture dependent path to the rClr native library

Description

Guess the directory where to expect the architecture dependent native library of a specified package e.g. for the package rClr, $R_HOME/library/rClr/libs/x64 This is a utility that is not specific to the rClr package

Usage

getNativeLibsPath(pkgName)

Arguments

pkgName
the name of a package, e.g. 'rClr'

Value

the prospective path in which a native library would be found, e.g. c:/R/library/rClr/libs/x64

[Package rClr version 0.9.1 Index]