mkClrObjRef {rClr}	R Documentation

Create if possible an S4 CLR object.

Description

Create if possible and adequate the S4 object that wraps the external pointer to a CLR object. Currently not exported, as this is unlikely to be recommended for use outside of unit tests and internal to rClr.

Usage

mkClrObjRef(obj, clrtype = NULL)

Arguments

obj
the presumed external pointer.

clrtype
character; the name of the CLR type for the object. If NULL, rClr retrieves the type name.

Value

a cobjRef S4 object if the argument is indeed an external pointer, otherwise returned unchanged if null or not an external pointer.

[Package rClr version 0.9.1 Index]