peekClrArgs {rClr}	R Documentation

Peek into the types of CLR objects arguments are converted to by rClr

Description

Advanced use only, to diagnose unexpected conditions in CLR method calls. Most users would not ever need it.

Usage

peekClrArgs(...)

Arguments

...
method arguments passed to .External

Value

a character message with type information about each argument.

Examples

## Not run: 
library(rClr)
peekClrArgs("a", numeric(0))

## End(Not run)

[Package rClr version 0.9.1 Index]