rClr-package {rClr}	R Documentation

R accessing .NET/Mono

Description

Accessing the Common Language Runtime (.NET/Mono) from R

Details

Package: 	 rClr

Type: 	 Package

Version: 	 0.8.3

Notes: 	 Works with R 3.5.x series. .NET code targets netstandard2.0.

Date: 	 2019-05-09

License: 	 LGPL 3

rClr is a package for low-level access from R to a Common Language Runtime (CLR). The supported CLR implementations are Microsoft '.NET' framework on Windows and Mono on several platforms, currently Windows and at least Debian Linux. rClr has been used in scientific modelling work since 2013-06. Running it on Mono has been relatively experimental, but as of November 2014 several factors have enabled very near functional parity with the Windows/.NET.

Version 	 Date 	 Notes

0.8.2: 	 2019-04-09 	 preview release - work on R 3.5.x, which has significant changes under the hood.

0.8.1: 	 2019-04-08 	 preview release - work on R 3.5.x, which has significant changes under the hood.

0.8.0: 	 2015-11-01 	 Tidy up build process and upgrade to latest recommended practices, notably with testthat.

0.7-9: 	 2015-08-30 	 allow compilation against boehm and SGen versions of mono lib. minor build improvements.

0.7-8: 	 2015-08-04 	 Update to use R.NET 1.6.5. Allow for backward compatibility down to R 2.15.3 - While all relevant tests work, note that the level of testing is not on par with tests on R 3.2.x.

0.7-7: 	 2015-06-18 	 Update to use R.NET 1.6.4. Functionally unchanged otherwise. Not a public release

0.7-6: 	 2015-05-07 	 update the R Markdown vignettes to work with the latest version of knitr. Not a public release

0.7-5: 	 2015-04-06 	 Internal release for work project. Compile against mono sgen rather than boehm.

I gratefully acknowledge Kosei Abe and others for their work on the dependency R.NET (see https://github.com/jmp75/rdotnet/graphs/contributors)

Author(s)

Jean-Michel Perraud jean-michel.perraud_at_csiro.au

[Package rClr version 0.9.1 Index]