setConvertAdvancedTypes {rClr}	R Documentation

Turn on/off the conversion of advanced data types with R.NET

Description

Turn on/off the conversion of advanced data types with R.NET. This will turn off the conversion of classes such as dictionaries into R lists, as these are not bidirectional and you may want to see and manipulate external pointers to dictionaries in some circumstances.

Usage

setConvertAdvancedTypes(enable = TRUE)

Arguments

enable
if true enable, otherwise disable

Examples

## Not run: 
library(rClr)
cTypename <- "Rclr.TestCases"
clrCallStatic(cTypename, "CreateStringDictionary")
setConvertAdvancedTypes(FALSE)
clrCallStatic(cTypename, "CreateStringDictionary")

## End(Not run)

[Package rClr version 0.9.1 Index]