setRDotNet {rClr}	R Documentation

Turn on/off R.NET

Description

Turn on or off the usage of the R.NET assemblies to convert CLR objects to R data structures. As of version 0.7.0, R.NET is the preferred way to convert data and is enabled by default.

Usage

setRDotNet(setit = TRUE)

Arguments

setit
if true enable, otherwise disable

Examples

## Not run: 
library(rClr)
## R.NET is currently used to convert complicated CLR types to sensible R equivalents
setRDotNet()
cTypename <- "Rclr.TestCases"
clrCallStatic(cTypename, "CreateStringDictionary")
setRDotNet(FALSE)
clrCallStatic(cTypename, "CreateStringDictionary")

## End(Not run)

[Package rClr version 0.9.1 Index]